developer_requirements.txt
ibm-watson >= 4.3.0
Markdown >= 3.2.1
ruamel.yaml >= 0.16.10